Fix compilation after recent main repo changes

+++ tracers/src/main/java/org/neo4j/io/pagecache/tracing/jfr/JfrCacheFlushEvent.java
@@ -64,7 +64,7 @@ public class JfrCacheFlushEvent extends TimedEvent implements MajorFlushEvent, F
     }
 
     @Override
-    public FlushEvent beginFlush( long filePageId, int cachePageId, PageSwapper swapper )
+    public FlushEvent beginFlush( long filePageId, long cachePageId, PageSwapper swapper )
     {
         flushes.getAndIncrement();
         JfrFlushEvent event = new JfrFlushEvent( bytesWritten );
